+++ src/playlist-controller.js
@@ -933,18 +933,12 @@ export class PlaylistController extends videojs.EventTarget {
 
     // Delete all buffered data to allow an immediate quality switch, then seek to give
     // the browser a kick to remove any cached frames from the previous rendtion (.04 seconds
-    // ahead is roughly the minimum that will accomplish this across a variety of content
+    // ahead was roughly the minimum that will accomplish this across a variety of content
     // in IE and Edge, but seeking in place is sufficient on all other browsers)
     // Edge/IE bug: https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/14600375/
     // Chrome bug: https://bugs.chromium.org/p/chromium/issues/detail?id=651904
     this.mainSegmentLoader_.resetEverything(() => {
-      // Since this is not a typical seek, we avoid the seekTo method which can cause segments
-      // from the previously enabled rendition to load before the new playlist has finished loading
-      if (videojs.browser.IE_VERSION || videojs.browser.IS_EDGE) {
-        this.tech_.setCurrentTime(this.tech_.currentTime() + 0.04);
-      } else {
-        this.tech_.setCurrentTime(this.tech_.currentTime());
-      }
+      this.tech_.setCurrentTime(this.tech_.currentTime());
     });
 
     // don't need to reset audio as it is reset when media changes
@@ -1000,19 +994,6 @@ export class PlaylistController extends videojs.EventTarget {
       if (!seekable.length) {
         // without a seekable range, the player cannot seek to begin buffering at the live
         // point
-        return false;
-      }
-
-      if (videojs.browser.IE_VERSION &&
-          this.tech_.readyState() === 0) {
-        // IE11 throws an InvalidStateError if you try to set currentTime while the
-        // readyState is 0, so it must be delayed until the tech fires loadedmetadata.
-        this.tech_.one('loadedmetadata', () => {
-          this.trigger('firstplay');
-          this.tech_.setCurrentTime(seekable.end(0));
-          this.hasPlayed_ = true;
-        });
-
         return false;
       }
 
@@ -1715,9 +1696,11 @@ export class PlaylistController extends videojs.EventTarget {
       audio: this.audioSegmentLoader_.getCurrentMediaInfo_() || {}
     };
 
+    const playlist = this.mainSegmentLoader_.getPendingSegmentPlaylist() || this.media();
+
     // set "main" media equal to video
     media.video = media.main;
-    const playlistCodecs = codecsForPlaylist(this.main(), this.media());
+    const playlistCodecs = codecsForPlaylist(this.main(), playlist);
     const codecs = {};
     const usingAudioLoader = !!this.mediaTypes_.AUDIO.activePlaylistLoader;
 
@@ -1738,7 +1721,7 @@ export class PlaylistController extends videojs.EventTarget {
     // no codecs, no playback.
     if (!codecs.audio && !codecs.video) {
       this.excludePlaylist({
-        playlistToExclude: this.media(),
+        playlistToExclude: playlist,
         error: { message: 'Could not determine codecs for playlist.' },
         playlistExclusionDuration: Infinity
       });
@@ -1763,13 +1746,13 @@ export class PlaylistController extends videojs.EventTarget {
       }
     });
 
-    if (usingAudioLoader && unsupportedAudio && this.media().attributes.AUDIO) {
-      const audioGroup = this.media().attributes.AUDIO;
+    if (usingAudioLoader && unsupportedAudio && playlist.attributes.AUDIO) {
+      const audioGroup = playlist.attributes.AUDIO;
 
       this.main().playlists.forEach(variant => {
         const variantAudioGroup = variant.attributes && variant.attributes.AUDIO;
 
-        if (variantAudioGroup === audioGroup && variant !== this.media()) {
+        if (variantAudioGroup === audioGroup && variant !== playlist) {
           variant.excludeUntil = Infinity;
         }
       });
@@ -1790,7 +1773,7 @@ export class PlaylistController extends videojs.EventTarget {
       }, '') + '.';
 
       this.excludePlaylist({
-        playlistToExclude: this.media(),
+        playlistToExclude: playlist,
         error: {
           internal: true,
           message
@@ -1817,7 +1800,7 @@ export class PlaylistController extends videojs.EventTarget {
 
       if (switchMessages.length) {
         this.excludePlaylist({
-          playlistToExclude: this.media(),
+          playlistToExclude: playlist,
           error: {
             message: `Codec switching not supported: ${switchMessages.join(', ')}.`,
             internal: true