+++ src/segment-loader.js
@@ -576,7 +576,7 @@ export default class SegmentLoader extends videojs.EventTarget {
     // TODO possibly move gopBuffer and timeMapping info to a separate controller
     this.gopBuffer_ = [];
     this.timeMapping_ = 0;
-    this.safeAppend_ = videojs.browser.IE_VERSION >= 11;
+    this.safeAppend_ = false;
     this.appendInitSegment_ = {
       audio: true,
       video: true
@@ -1163,6 +1163,7 @@ export default class SegmentLoader extends videojs.EventTarget {
    */
   resetEverything(done) {
     this.ended_ = false;
+    this.activeInitSegmentId_ = null;
     this.appendInitSegment_ = {
       audio: true,
       video: true
@@ -1981,6 +1982,10 @@ export default class SegmentLoader extends videojs.EventTarget {
 
   getMediaInfo_(segmentInfo = this.pendingSegment_) {
     return this.getCurrentMediaInfo_(segmentInfo) || this.startingMediaInfo_;
+  }
+
+  getPendingSegmentPlaylist() {
+    return this.pendingSegment_ ? this.pendingSegment_.playlist : null;
   }
 
   hasEnoughInfoToAppend_() {