+++ test/loader-common.js
@@ -1,5 +1,4 @@
 import QUnit from 'qunit';
-import videojs from 'video.js';
 import xhrFactory from '../src/xhr';
 import Config from '../src/config';
 import document from 'global/document';
@@ -941,16 +940,7 @@ export const LoaderCommonFactory = ({
 
     // only main/fmp4 segment loaders use async appends and parts/partIndex
     if (usesAsyncAppends) {
-      let testFn = 'test';
-
-      if (videojs.browser.IE_VERSION) {
-        testFn = 'skip';
-      }
-
-      // this test has a race condition on ie 11 that causes it to fail some of the time.
-      // Since IE 11 isn't really a priority and it only fails some of the time we decided to
-      // skip this on IE 11.
-      QUnit[testFn]('playlist change before any appends does not error', function(assert) {
+      QUnit.test('playlist change before any appends does not error', function(assert) {
         return this.setupMediaSource(loader.mediaSource_, loader.sourceUpdater_).then(() => {
           loader.playlist(playlistWithDuration(50, {
             uri: 'bar-720.m3u8',