+++ test/util/text-tracks.test.js
@@ -123,3 +123,38 @@ test('creates cues for timed metadata', function(assert) {
     'added one metadata cues'
   );
 });
+
+test('does nothing if frames are undefined', function(assert) {
+  addMetadata({
+    inbandTextTracks: this.inbandTextTracks,
+    timestampOffset: this.timestampOffset,
+    videoDuration: 1,
+    metadataArray: [{
+      cueTime: 1
+    }]
+  });
+
+  assert.strictEqual(
+    this.inbandTextTracks.metadataTrack_.cues.length,
+    0,
+    'added no metadata cues'
+  );
+});
+
+test('does nothing if frames.length is 0', function(assert) {
+  addMetadata({
+    inbandTextTracks: this.inbandTextTracks,
+    timestampOffset: this.timestampOffset,
+    videoDuration: 1,
+    metadataArray: [{
+      cueTime: 1,
+      frames: []
+    }]
+  });
+
+  assert.strictEqual(
+    this.inbandTextTracks.metadataTrack_.cues.length,
+    0,
+    'added no metadata cues'
+  );
+});